% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coords-class.R
\docType{class}
\name{Coords class}
\alias{Coords class}
\alias{Coords-class}
\title{Coords class}
\description{
This is a class union, providing a convenient class for
 coordinates in several formats.
}
\details{
The class accepts coordinates in any of the following formats:
 1. SpatialPoints
 2. two-columned matrix
 3. vector of length 2
}
\examples{
showClass("Coords")
}
\keyword{classes}
