\name{Coords class}
\docType{class}
\alias{Coords-class}

\title{Coords class}
\description{
This is a class union, providing a convenient class for coordinates in several formats.
The class accepts coordinates in any of the following formats:
1. SpatialPoints
2. two-columned matrix
3. vector of length 2
}

\author{ Jacob van Etten <jacobvanetten@yahoo.com> }
\examples{
showClass("Coords")

}
\keyword{classes}
