% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDist.R
\docType{data}
\name{genDist}
\alias{genDist}
\alias{popCoord}
\title{Genetic distances and coordinates of haplogroup R1b1b2 populations in Europe}
\format{
dist object (genDist) and dataframe (popCoord)
}
\source{
Distances calculated from Table S1 in 
Balaresque et al. (2010). 
Coordinates from Table 1.
}
\description{
genDist: Genetic distances between 23 populations of Y-chromosome 
haplogroup R1b1b2 in Europe, calculated from haplotype microsatellite 
data (see source).
}
\details{
The distance measure used is the negative logarithm of the shared 
proportion of alleles.

popCoord: Geographic coordinates of the same populations.
}
\references{
Balaresque P., et al. 2010. A predominantly Neolithic 
origin for European paternal lineages. 
\emph{PLoS Biology} 8(1): e1000285.
}
\keyword{datasets}
