\name{overlap}
\alias{nonoverlap}
\alias{overlap}

\title{Overlap and nonoverlap of trajectories}
\description{
Special functions to calculate the degree of overlap and nonoverlap between trajectories}
\usage{
overlap(a, b)
nonoverlap(a, b)
}

\arguments{
  \item{a}{ \code{matrix} object}
  \item{b}{ \code{matrix} object}  
}
\details{
These functions are used by the \code{pathInc()} as defaults. 
}
\value{
matrix
}
\note{
The functions are provided here to assist the user in defining alternative measures of overlap / nonoverlap.
}
\author{ Jacob van Etten \email{jacobvanetten@yahoo.com} }
\keyword{ spatial }

