\name{[-methods}
\docType{methods}
\alias{[,Transition,index,missing,missing-method}
\alias{[,Transition,index,index,missing-method}
\alias{[<-,Transition,index,missing,dsCMatrix-method}
\alias{[<-,Transition,index,index,ANY-method}
\alias{transitionMatrix,Transition-method}
\alias{transitionMatrix<-,Transition,dsCMatrix-method}
\alias{transitionCells,Transition-method}

\title{ Extracting and replacing: class Transition}
\description{
Methods for function \code{[} in Package `base' ~~
}
\section{Methods}{
\describe{
\item{x = "Transition", i = "index", j = "missing", drop = "missing"}{ ~~describe this method here }
\item{x = "Transition", i = "index", j = "missing", value = "ANY"}{ ~~describe this method here }
\item{transition = "Transition"}{ ~~describe this method here }
\item{transition = "Transition", value = "dsCMatrix"}{ ~~describe this method here }
\item{transition = "Transition"}{ ~~describe this method here }
}}
\keyword{methods}