\name{Transition-extract-replace2}
\alias{transitionMatrix}
\alias{matrixValues}
\alias{transitionCells}
\alias{transitionMatrix,Transition-method}
\alias{transitionMatrix<-,Transition,dsCMatrix-method}
\alias{transitionCells,Transition-method}
\alias{matrixValues,Transition-method}
\alias{matrixValues<-,Transition,character-method}
\title{Extract or change elements of Transition objects}
\usage{
transitionMatrix(transition)
transitionCells(transition)
matrixValues(transition)
}