% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdistance.R
\docType{package}
\name{gdistance}
\alias{gdistance}
\alias{gdistance-package}
\title{Distances and Routes on Geographical Grids}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Calculate distances and routes on geographic grids.
}
\seealso{
\strong{Useful links:}
\itemize{
\item{Development repository: 
 \url{https://github.com/agrobioinfoservices/gdistance}}
\item{Static documentation: 
 \url{https://github.com/agrobioinfoservices/gdistance}}
\item{Report bugs: 
 \url{https://github.com/agrobioinfoservices/gdistance/issues}}
}
}
\author{
Jacob van Etten and Kauê de Sousa
}
