% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdistance.R
\docType{package}
\name{gdistance}
\alias{gdistance}
\alias{gdistance-package}
\title{Distances and Routes on Geographical Grids}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides classes and functions to calculate various distance measures and routes in heterogeneous geographic spaces represented as grids. The package implements measures to model dispersal histories first presented by van Etten and Hijmans (2010) \doi{10.1371/journal.pone.0012060}. Least-cost distances as well as more complex distances based on (constrained) random walks can be calculated. The distances implemented in the package are used in geographical genetics, accessibility indicators, and may also have applications in other fields of geospatial analysis.
}
\seealso{
\strong{Useful links:}
\itemize{
\item{Package's paper: 
 \doi{https://doi.org/10.18637/jss.v076.i13}}
\item{Development repository: 
 \url{https://github.com/AgrDataSci/gdistance}}
\item{Static documentation: 
 \url{https://github.com/AgrDataSci/gdistance}}
\item{Report bugs: 
 \url{https://github.com/AgrDataSci/gdistance/issues}}
}
}
\author{
Jacob van Etten and Kauê de Sousa
}
