% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathInc.R
\name{overlap}
\alias{overlap}
\alias{nonoverlap}
\title{Overlap and nonoverlap of trajectories}
\usage{
overlap(a, b)
}
\arguments{
\item{a}{\code{matrix} object}

\item{b}{\code{matrix} object}
}
\value{
A matrix object
}
\description{
Special functions to calculate the degree of overlap and
 nonoverlap between trajectories
}
\details{
These functions are used by the \code{pathInc()} as defaults.
}
\note{
The functions are provided here to assist the user in defining
 alternative measures of overlap / nonoverlap.
}
\author{
Jacob van Etten
}
\keyword{spatial}
