\name{Transition methods}
\docType{methods}
\alias{[,Transition,numeric,numeric,missing-method}
\alias{[,Transition,matrix,missing,missing-method}
\alias{[<-,Transition,matrix,missing,ANY-method}
\alias{[<-,Transition,numeric,numeric,ANY-method}
\alias{==,Transition,Transition-method}
\alias{coerce,RasterLayer,Transition-method}
\alias{coerce,Transition,RasterLayer-method}

\title{ Extracting and replacing: class Transition}
\description{
Methods for functions \code{[} and \code{[<-} for object of the class Transition
Methods for functions \code{transitionMatrix} and \code{transitionMatrix<-}
}
\keyword{methods}