\name{Transition* methods}
\docType{methods}
\alias{[,TransitionLayer,numeric,numeric,missing-method}
\alias{[,TransitionLayer,matrix,missing,missing-method}
\alias{[<-,TransitionLayer,matrix,missing,ANY-method}
\alias{[<-,TransitionLayer,numeric,numeric,ANY-method}
\alias{[[,TransitionStack,numeric,missing-method}
\alias{[[<-,TransitionStack,numeric,missing,TransitionData-method}

\title{ Extracting and replacing: class Transition}
\description{
Methods for functions \code{[} and \code{[<-} for object of the class TransitionLayer.
Methods for functions \code{[[} and \code{[[<-} for object of the class TransitionStack.
}
\keyword{methods}