\name{Transition methods}
\docType{methods}
\alias{[,TransitionLayer,numeric,numeric,missing-method}
\alias{[,TransitionLayer,matrix,missing,missing-method}
\alias{[<-,TransitionLayer,matrix,missing,ANY-method}
\alias{[<-,TransitionLayer,numeric,numeric,ANY-method}


\title{ Extracting and replacing: class Transition}
\description{
Methods for functions \code{[} and \code{[<-} for object of the class Transition*
Methods for functions \code{transitionMatrix} and \code{transitionMatrix<-}
}
\keyword{methods}