% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transition-slots.R
\docType{methods}
\name{Transition-slots}
\alias{Transition-slots}
\alias{transitionMatrix}
\alias{nlayers,TransitionLayer-method}
\alias{nlayers,TransitionStack-method}
\alias{transitionMatrix,TransitionLayer,logical-method}
\alias{transitionMatrix,TransitionLayer,missing-method}
\alias{transitionMatrix,TransitionData,missing-method}
\alias{matrixValues}
\alias{matrixValues<-}
\alias{transitionCells}
\alias{transitionData}
\alias{transitionCells,TransitionLayer-method}
\alias{transitionCells,TransitionData-method}
\alias{matrixValues,TransitionLayer-method}
\alias{matrixValues<-,TransitionLayer,character-method}
\alias{matrixValues,TransitionStack-method}
\alias{matrixValues<-,TransitionStack,character-method}
\alias{transitionData,TransitionLayer-method}
\alias{transitionData,TransitionStack-method}
\title{Extract or change elements of Transition* objects}
\usage{
transitionMatrix(x, inflate)
}
\arguments{
\item{x}{object of class \code{Transition*}}

\item{inflate}{logical (default is \code{TRUE})}
}
\description{
These functions are to be used to access slots of Transition* objects.
}
\keyword{methods}
