\name{Transition slots}
\alias{transitionMatrix}
\alias{transitionMatrix<-}
\alias{matrixValues}
\alias{matrixValues<-}
\alias{transitionCells}
\alias{transitionMatrix,Transition-method}
\alias{transitionMatrix<-,Transition,sparseMatrix-method}
\alias{transitionCells,Transition-method}
\alias{matrixValues,Transition-method}
\alias{matrixValues<-,Transition,character-method}
\title{Extract or change elements of Transition objects}
\description{
These functions are to be used to access slots of Transition objects directly.
}
\usage{
transitionMatrix(transition)
transitionCells(transition)
matrixValues(transition)
}
\arguments{
  \item{transition}{object of class Transition}
}